% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_roclist.R
\name{roclist}
\alias{roclist}
\title{roclist
this function makes ROC analyses}
\usage{
roclist(kras,geneset,outcome="os.status")
}
\description{
roclist
this function makes ROC analyses
}
\examples{
data(kras)
kras<-rocpreprocess(kras)
data(geneset)
roc.list<-roclist(kras,geneset,outcome="os.status")
roc.list
}
