% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_rocplot.r
\name{rocplot}
\alias{rocplot}
\title{rocplot
this function makes the ROC plot}
\usage{
rocplot(roc.list,line=0.5,title="KRAS network ~ OS pancreas cancer",police=14)
}
\description{
rocplot
this function makes the ROC plot
}
\examples{
data(kras)
kras<-rocpreprocess(kras)
data(geneset)
roc.list<-roclist(kras,geneset,outcome="os.status")
rocplot(roc.list,line=0.5,title="KRAS network ~ OS pancreas cancer",police=14)
}
