% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_rocpreprocess.r
\name{rocpreprocess}
\alias{rocpreprocess}
\title{rocpreprocess
this function takes input dataframe and remove missing data}
\usage{
rocpreprocess(kras)
}
\description{
rocpreprocess
this function takes input dataframe and remove missing data
}
\examples{
data(kras)
kras<-rocpreprocess(kras)
head(kras)
}
