% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_roctable.r
\name{roctable}
\alias{roctable}
\title{roctable
this function output a table of AUC results}
\usage{
roctable(roc.list)
}
\description{
roctable
this function output a table of AUC results
}
\examples{
data(kras)
kras<-rocpreprocess(kras)
data(geneset)
roc.list<-roclist(kras,geneset,outcome="os.status")
roc.table<-roctable(roc.list)
roc.table
}
